split nativecontract
